Rename card file Move debug function to tools file move credential file
